// apps/api/src/modules/inventory/entities/item-photo.entity.d.ts
import { InventoryItem } from './inventory-item.entity';
export declare class ItemPhoto {
    id: string;
    item: InventoryItem;
    itemId: string;
    s3Key: string;
    cdnUrl: string;
    sortOrder: number;
    isPrimary: boolean;
    createdAt: Date;
}
//# sourceMappingURL=item-photo.entity.d.ts.map