// apps/api/src/modules/inventory/dto/update-inventory-item.dto.d.ts
import { CreateInventoryItemDto } from './create-inventory-item.dto';
declare const UpdateInventoryItemDto_base: import("@nestjs/common").Type<Partial<CreateInventoryItemDto>>;
export declare class UpdateInventoryItemDto extends UpdateInventoryItemDto_base {
    status: string;
}
export {};
//# sourceMappingURL=update-inventory-item.dto.d.ts.map