// apps/api/src/modules/inventory/dto/query-inventory.dto.d.ts
export declare class QueryInventoryDto {
    page: number;
    limit: number;
    condition: string;
    status: string;
    brandId: string;
    modelId: string;
    branchId: string;
    minPrice: number;
    maxPrice: number;
    search: string;
}
//# sourceMappingURL=query-inventory.dto.d.ts.map